kapa-pdf-takehome_habibou_sissoko/requirements.txt
streamlit==1.45.1
PyMuPDF==1.26.0
lancedb==0.22.1
pyarrow==20.0.0
tantivy==0.24.0
pylance==0.27.2
openai==1.82.0
mistune==3.0.1
langchain==0.1.17
pymupdf4llm==0.0.24
black==25.1.0
isort==6.0.1

numpy==1.26.4
pytesseract==0.3.13
Pillow==11.2.1
torch==2.7.1
doclayout-yolo==0.0.4
transformers==4.52.4
timm==1.0.15

python-doctr[torch]